dom/media/tests/mochitest/test_peerConnection_bug1042791.html

<!DOCTYPE HTML>
<html>
<head>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="head.js"></script>
  <script type="application/javascript" src="mediaStreamPlayback.js"></script>
  <script type="application/javascript" src="pc.js"></script>
  <script type="application/javascript" src="templates.js"></script>
  <script type="application/javascript" src="turnConfig.js"></script>
</head>
<body>
<pre id="test">
<script type="application/javascript;version=1.8">
  createHTML({
    bug: "1040346",
    title: "Basic H.264 GMP video-only peer connection"
  });

  var test;
  runNetworkTest(function (options) {
    options = options || { };
    options.h264 = true;
    test = new PeerConnectionTest(options);
    test.setMediaConstraints([{video: true}], [{video: true}]);
    test.chain.removeAfter("PC_LOCAL_CREATE_OFFER");

    test.chain.append([[
      "PC_LOCAL_VERIFY_H264_OFFER",
      function (test) {
        ok(!test.pcLocal._last_offer.sdp.contains("profile-level-id=0x42e00c"),
          "H264 offer does not contain profile-level-id=0x42e00c");
        ok(test.pcLocal._last_offer.sdp.contains("profile-level-id=42e00c"),
          "H264 offer contains profile-level-id=42e00c");
        test.next();
      }
    ]]);

    test.run();
  });
</script>
</pre>
</body>
</html>
